+++ kusitms5/src/main/java/com/kusitms/kusitms5/dto/reviewDto.java
@@ -14,7 +14,7 @@ public class reviewDto {
 
     String reviewUserName;
     String reviewMemo;
-    int reviewScore;
+    double reviewScore;
 
     public reviewDto(Review review){
         reviewUserName = review.getUser().getNickname();
